Move src/tests.rs to src/functional_tests.rs
